requirements.txt
mediapipe==0.10.14
opencv-python==4.10.0.84
ultralytics==8.2.0
catboost==1.2.5
xgboost==2.0.3
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
pyyaml==6.0.1

